scale: remove dest globals and coord parameters
Remove the dest_buffer and dest_pitch globals from i_scale and
refactor all dependent code to receive appropriate arguments
instead. This is mostly the scalers themselves.

Remove the coordinate arguments to scalers (x1,x2,y1,y2). The
scalers are only called with the full internal screen as the
source to scale from (0,0,320,200). Half the scalers only
supported this anyway, and had guard code in to protect against
being called with other arguments. Remove the redundant guard
code too. Minor refactoring of scalers to simplify things with
the constants (mostly term removal with a 0-product).

Finally follow the coordinates up the call chain to BlitArea,
which is only ever called for full-(internal)-screen update and
gets renamed to "Blit".