WIP: add SetDefaultSaveName
SetDefaultSaveName which sets the default name for a save slot
when selected if there isn't one already. Vanilla behaviour is
to have a blank name. Construct a more descriptive default from
the current game variant and map number.

The intention is to only execute this behaviour if the save slot
was selected via a gamepad (which would suggest the player can't
conveniently enter a save description themselves).