SetDefaultSaveName: ditch gamedescription, +skill
Stop looking at gamedescription which is very unlikely to be short
enough to fit into the save slot anyway.

If we can fit the skill level into the slot (as "skillX"), do so.

TODO: rework the string length checks. The array we're writing into
isn't (or doesn't have to be) null-terminated, but things like
M_snprintf work on that basis.